Added shell option to change EOI string
I had an issue with the string 'EOI' also occurring in my PS output when invoking PS commands, which caused truncated output. 

To solve this issue, I added an option to the Shell instantiation to change the EOI string if necessary.

You'll see I also added a regular expression to replace characters that need escaping in PowerShell (e.g. my EOI string is now "`<EOI`>" with backticks to escape the < and >) to detect the end of invoke, because these special characters will occur less likely in console output.